Utilise is_transient instead of recordToDB
Updated the `sparkplugNode.ts` script to strip `recordToDB` property from the metric after it has fulfilled its purpose by carrying intentions to inform `is_transient`. This step is undertaken to conserve memory space and avoid potential confusion as the left-over property in the payload would be superfluous and could lead to misunderstandings. A new constant `isTransient` has been defined to facilitate this exclusion and is used to set the is_transient property on the Sparkplug metric.