Fix potential null reference in sparkplugNode
This commit addresses a potential null reference error on property `recordToDB`. Previously, the check for the `recordToDB` was assuming that it's always present which can lead to errors when metric does not contain such property. Now we add another optional chaining before `.value` to make sure that it's not undefined or null before checking its value. This aims to improve the overall robustness of the code.