Remove Typescript.
This is just the result of running `tsc` with some keep-your-fingers-off
options in tsconfig.json. If we are going to do this it would need
testing properly to make sure it still works and then going through to
remove some of the unnecessary changes. The blank lines, in particular,
want to come back.