Hotfix/s7 Device Read Config Overwrite Fix (#39)
Fix an issue with 2 devices sharing the same connection.

When the second device is configured it clears the `s7Vars` of the first
device.

This fix adds second config to the to a variable list instead of
clearing it.

(cherry picked from commit 78aa216d1b230047b78c258f774e54071ce16c34)